physx fixes for macOS

+++ deps/physx/physx/source/foundation/FdTempAllocator.cpp
@@ -64,7 +64,7 @@ void* PxTempAllocator::allocate(size_t size, const char* filename, PxI32 line)
 	if(!size)
 		return 0;
 
-	PxU32 index = PxMax(PxHighestSetBit(PxU32(size) + sizeof(Chunk) - 1), sMinIndex);
+	PxU32 index = PxMax(PxHighestSetBit(uint32_t(PxU32(size) + sizeof(Chunk) - 1)), sMinIndex);
 
 	Chunk* chunk = 0;
 	if(index < sMaxIndex)
